feat: gera erros e adiciona verificações

+++ src/domains/User/controller/index.ts
@@ -1,20 +1,21 @@
 import UserService from '../services/UserService';
 import { Router, Request, Response, NextFunction } from 'express';
-// import { loginMiddleware, verifyJWT } from '../../../middlewares/auth-middlewares';
-// import UserRoles from '../../../../utils/constants/userRoles';
-// import statusCodes from '../../../../utils/constants/statusCodes';
-
+import UserRoles from '../../../../utils/constants/userRoles';
+import statusCodes from '../../../../utils/constants/statusCodes';
+import checkRoles from '../../../middlewares/checkRole';
+import { verifyJWT } from '../../../middlewares/auth-middlewares';
+import { NotAuthorizedError } from '../../../../errors/NotAuthorizedError';
+import app from '../../../../config/expressConfig';
 // import { userInfo } from 'os';
 // import { parse } from 'path';
-
-import statusCodes from '../../../../utils/constants/statusCodes';
 
 const router = Router();
 
 // router.post('/login', async (req: Request, res: Response, next: NextFunction) => {});
 // router.post('/logout', async (req: Request, res: Response, next: NextFunction) => {});
 
-router.post('/create', async (req: Request, res: Response, next: NextFunction) => {
+app.use(verifyJWT); // deve ficar abaixo de rota de login
+router.post('/create', checkRoles([UserRoles.ADMIN, UserRoles.USER]), async (req: Request, res: Response, next: NextFunction) => {
 	try {
 		await UserService.create(req.body);
 		res.status(statusCodes.CREATED).json('Usuário criado com sucesso!');
@@ -43,6 +44,9 @@ router.get('/email/:email', async (req: Request, res: Response, next: NextFuncti
 
 router.get('/allUsers', async (req: Request, res: Response, next: NextFunction) => {
 	try {
+		if (req.user.role != UserRoles.ADMIN) {
+			throw new NotAuthorizedError('Você não tem permissão para visualizar todos os usuários.');
+		}
 		const users = await UserService.read();
 		res.status(statusCodes.SUCCESS).json(users);
 	} catch (error) {
@@ -52,6 +56,10 @@ router.get('/allUsers', async (req: Request, res: Response, next: NextFunction) 
 
 router.delete('/delete/:id', async (req: Request, res: Response, next: NextFunction) => {
 	try {
+		if (req.user.role == UserRoles.USER && req.user.id != parseInt(req.params.id)) {
+			throw new NotAuthorizedError('Você não tem permissão para deletar outro usuário.');
+		}
+
 		await UserService.deleteUser(parseInt(req.params.id));
 		res.status(statusCodes.NO_CONTENT).json('Usuário deletado com sucesso!');
 	}
@@ -62,10 +70,11 @@ router.delete('/delete/:id', async (req: Request, res: Response, next: NextFunct
 
 router.put('/update', async (req: Request, res: Response, next: NextFunction) => {
 	try {
-		const { id, name, email, photo, password, role } = req.body;
-		if (req.body.id === undefined) {
-			res.status(statusCodes.BAD_REQUEST).json('O campo id é obrigatório!');
+		if (req.user.id != parseInt(req.body.id)) {
+			throw new NotAuthorizedError('Você não tem permissão para atualizar outro usuário.');
 		}
+
+		const { id, name, email, photo, password, role } = req.body;
 		await UserService.updateUser({ id: parseInt(id), name, email, photo, password, role });
 		res.status(statusCodes.NO_CONTENT).json('Usuário atualizado com sucesso!');
 	} catch (error) {
